fix typo

+++ src/input.hpp
@@ -93,7 +93,7 @@ T Input::Get(std::string blockName, std::string paramName, int num) {
     // The following mess with pointers is required since we do not have access to constexpr if
     // in c++ 14, hence we need to cast T to all of the available type we support.
     if(typeid(T) == typeid(int)) {
-      double *v = reinterpret_cast<int*>( &value);
+      int *v = reinterpret_cast<int*>( &value);
       double dv = std::stod(paramString, NULL);
       int iv = static_cast<int>(std::round(dv));
       if (std::abs((dv - iv)/dv) > 1e-14) {
